Move app files to separate folder
