+++ backend/app/app.js
@@ -1,0 +1,5 @@
+const express = require('express');
+
+const app = express();
+
+module.exports = app;
