+++ backend/app/cors.js
@@ -1,0 +1,5 @@
+const cors = require('cors');
+
+const app = require('./app');
+
+app.use(cors());
